+++ tests/scripts/test_send_targets.py
@@ -17,7 +17,6 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 import unittest
-import os
 from unittest.mock import patch
 from pathlib import Path
 from argparse import Namespace
@@ -25,18 +24,18 @@ from lxml import etree
 from . import GmpMockFactory, load_script
 
 
-CWD = os.path.abspath(os.path.join(__file__, '../'))
+CWD = Path(__file__).absolute().parent
 
 
 class SendTargetTestCase(unittest.TestCase):
     def setUp(self):
         self.send_targets = load_script(
-            os.path.join(CWD, '../../scripts'), 'send-targets'
+            (CWD.parent.parent / 'scripts'), 'send-targets'
         )
 
     @patch('gvm.protocols.latest.Gmp', new_callable=GmpMockFactory)
     def test_sent_target(self, mock_gmp: GmpMockFactory):
-        target_xml_path = Path(CWD, 'example_target.xml')
+        target_xml_path = CWD / 'example_target.xml'
         target_xml_str = target_xml_path.read_text()
 
         mock_gmp.mock_response(
@@ -65,7 +64,7 @@ class SendTargetTestCase(unittest.TestCase):
     @patch('builtins.input', lambda *args: 'n')
     @patch('gvm.protocols.latest.Gmp', new_callable=GmpMockFactory)
     def test_sent_target_no_credential(self, mock_gmp: GmpMockFactory):
-        target_xml_path = Path(CWD, 'example_target.xml')
+        target_xml_path = CWD / 'example_target.xml'
         target_xml_str = target_xml_path.read_text()
 
         mock_gmp.mock_response(
@@ -97,40 +96,3 @@ class SendTargetTestCase(unittest.TestCase):
 
         with self.assertRaises(SystemExit):
             self.send_targets.check_args(args2)
-
-    @patch('builtins.input', lambda *args: 'y')
-    def test_yes(self):
-        yes = self.send_targets.yes_or_no('foo?')
-        self.assertTrue(yes)
-
-    @patch('builtins.input', lambda *args: 'n')
-    def test_no(self):
-        no = self.send_targets.yes_or_no('bar?')
-        self.assertFalse(no)
-
-    def test_error_and_exit(self):
-        with self.assertRaises(SystemExit):
-            self.send_targets.error_and_exit('foo')
-
-    def test_create_xml_tree(self):
-        target_xml_path = Path(CWD, 'example_target.xml')
-
-        tree = self.send_targets.create_xml_tree(str(target_xml_path))
-        self.assertIsInstance(
-            tree, etree._Element  # pylint: disable=protected-access
-        )
-        self.assertEqual(tree.tag, 'get_targets_response')
-
-    def test_create_xml_tree_invalid_file(self):
-        target_xml_path = Path(CWD, 'invalid_file.xml')
-
-        with self.assertRaises(SystemExit):
-            with self.assertRaises(OSError):
-                self.send_targets.create_xml_tree(str(target_xml_path))
-
-    def test_create_xml_tree_invalid_xml(self):
-        target_xml_path = Path(CWD, 'invalid_xml.xml')
-
-        with self.assertRaises(SystemExit):
-            with self.assertRaises(etree.Error):
-                self.send_targets.create_xml_tree(str(target_xml_path))