+++ CHANGELOG.md
@@ -19,7 +19,9 @@ $ cd gvm-tools && git log
 * Added `pretty_print` to `pyshell` by default, so it does not need to be manually imported [#305](https://github.com/greenbone/gvm-tools/pull/305)
 * Added tests for `helper` module [#310](https://github.com/greenbone/gvm-tools/pull/310)
 * Added tests for `parser` module [#311](https://github.com/greenbone/gvm-tools/pull/311)
-* Added test for `script/send-target.gmp.py` [#314](https://github.com/greenbone/gvm-tools/pull/314/)
+* Added tests for `script/send-target.gmp.py` [#314](https://github.com/greenbone/gvm-tools/pull/314/)
+* Added tests for `script/send-tasks.gmp.py` [#317](https://github.com/greenbone/gvm-tools/pull/317)
+* Adding useful script helper functions to the `helper.py` [#317](https://github.com/greenbone/gvm-tools/pull/317)
 
 ### Changed
 ### Deprecated